Testing the hAPI composition
TO DO: The communication block will be absorbed into an abstract board object which will be the base class for the hapkit and haply boards. 

TO DO: Right now the actuator objects are instantiated when the device object is created. I want to adapt this so that when each device is created it instantiates a specific device object that has the number of actuators and the specific parameters of that device. 

TO DO: Create the appropriated enumerated types for the hAPI and hAPI convenience APIs